+++ vllm/model_executor/layers/quantization/quark/quark_moe.py
@@ -1185,7 +1185,7 @@ class QuarkOCP_MX_MoEMethod_OSS(QuarkOCP_MX_MoEMethod):
                 "EPLB not supported for `QuarkW4MXFp4MoEMethod_OSS` yet."
             )
 
-        from vllm.model_executor.layers.fused_moe.gpt_oss_triton_kernels_moe import (  # noqa: E501
+        from vllm.model_executor.layers.fused_moe.experts.gpt_oss_triton_kernels_moe import (  # noqa: E501
             triton_kernel_moe_forward,
         )
